third_party/WebKit/LayoutTests/platform/mac-mac10.10/virtual/outofblink-cors/external/wpt/XMLHttpRequest/open-url-redirected-worker-origin-expected.txt

This is a testharness.js-based test.
PASS XMLHttpRequest: redirected worker scripts, origin and referrer 
FAIL Referer header assert_equals: expected "Referer: http://web-platform.test:8001/XMLHttpRequest/resources/workerxhr-origin-referrer.js\n" but got "Referer: http://web-platform.test:8001/XMLHttpRequest/resources/redirect.py?location=workerxhr-origin-referrer.js\n"
PASS Request URL test 
PASS Origin header 
Harness: the test ran to completion.
